+++ my-app/src/components/CardForm/CardPasswordSection/CardPasswordSection.tsx
@@ -1,7 +1,6 @@
-import { useState } from 'react';
 import PasswordInput from './PasswordInput/PasswordInput';
-import { getCardPasswordError } from '../../../utils/validation';
 import CommonSection from '../../../common/CommonSection/CommonSection';
+import usePassword from '../../../hooks/usePassword';
 
 interface Props {
   value: string;
@@ -9,25 +8,10 @@ interface Props {
 }
 
 function CardPasswordSection({ value, updateValue }: Props) {
-  const [error, setError] = useState<boolean>(false);
-  const [errorMessage, setErrorMessage] = useState('');
-
-  function handleOnChange(inputValue: string) {
-    if (!/^[0-9]*$/.test(inputValue)) {
-      setErrorMessage('비밀번호는 숫자만 입력 가능합니다.');
-      return;
-    }
-
-    updateValue(inputValue);
-    setErrorMessage('');
-  }
-
-  function handleOnBlur() {
-    const validation = getCardPasswordError(value);
-
-    setError(validation.error);
-    setErrorMessage(validation.message);
-  }
+  const { error, errorMessage, handleOnChange, handleOnBlur } = usePassword({
+    value,
+    updateValue,
+  });
 
   return (
     <CommonSection
