+++ my-app/src/components/CardForm/ExpirationDateSection/ExpirationDateSection.tsx
@@ -1,7 +1,6 @@
-import { useEffect, useRef, useState } from 'react';
-import { getCardMonthError, getCardYearError } from '../../../utils/validation';
 import CommonSection from '../../../common/CommonSection/CommonSection';
 import NumberInput from '../../../common/NumberInput/NumberInput';
+import useExpirationDate from '../../../hooks/useExpirationDate';
 
 interface Props {
   value: {
@@ -12,60 +11,14 @@ interface Props {
 }
 
 export default function ExpirationDateSection({ value, updateValue }: Props) {
-  const monthInputRef = useRef<HTMLInputElement>(null);
-  const yearInputRef = useRef<HTMLInputElement>(null);
-  const [errors, setErrors] = useState({ month: false, year: false });
-  const [errorMessage, setErrorMessage] = useState('');
-
-  useEffect(() => {
-    const animationFrameId = requestAnimationFrame(() => {
-      monthInputRef.current?.focus();
-    });
-
-    return () => {
-      cancelAnimationFrame(animationFrameId);
-    };
-  }, []);
-
-  function handleOnChange(inputValue: string, type: 'month' | 'year') {
-    if (!/^[0-9]*$/.test(inputValue)) {
-      setErrorMessage('유효기간은 숫자만 입력 가능합니다.');
-      return;
-    }
-
-    const newValue = { ...value, [type]: inputValue };
-    updateValue(newValue);
-    setErrorMessage('');
-
-    if (type === 'month' && inputValue.length === 2) {
-      requestAnimationFrame(() => {
-        yearInputRef.current?.focus();
-      });
-    }
-  }
-
-  function handleOnBlur(type: 'month' | 'year') {
-    const validation =
-      type === 'month'
-        ? getCardMonthError(value.month)
-        : getCardYearError(value.year);
-
-    const otherValidation =
-      type === 'month'
-        ? getCardYearError(value.year)
-        : getCardMonthError(value.month);
-
-    setErrors({
-      ...errors,
-      [type]: validation.error,
-    });
-
-    setErrorMessage(
-      type === 'month'
-        ? validation.message
-        : validation.message || otherValidation.message,
-    );
-  }
+  const {
+    errors,
+    errorMessage,
+    handleOnChange,
+    handleOnBlur,
+    monthInputRef,
+    yearInputRef,
+  } = useExpirationDate({ value, updateValue });
 
   return (
     <CommonSection
